requirements.txt
requests
lxml
PyYAML
beautifulsoup4
